change default skip_value in gtx.as_connectivity
