+++ src/gt4py/_core/definitions.py
@@ -452,6 +452,8 @@ class NDArrayObject(Protocol):
 
     def astype(self, dtype: npt.DTypeLike) -> NDArrayObject: ...
 
+    def any(self) -> bool: ...
+
     def __getitem__(self, item: Any) -> NDArrayObject: ...
 
     def __abs__(self) -> NDArrayObject: ...
