+++ tests/next_tests/unit_tests/conftest.py
@@ -15,6 +15,7 @@ import pytest
 
 import gt4py.next as gtx
 from gt4py.next import backend, common
+from gt4py.next.embedded import nd_array_field
 from gt4py.next.iterator import runtime
 from gt4py.next.program_processors import program_formatter
 
@@ -101,3 +102,16 @@ class DummyConnectivity(common.Connectivity):
     has_skip_values: int
     source_dim: gtx.Dimension = gtx.Dimension("dummy_origin")
     codomain: gtx.Dimension = gtx.Dimension("dummy_neighbor")
+
+
+def nd_array_implementation_params():
+    for xp in nd_array_field._nd_array_implementations:
+        if hasattr(nd_array_field, "cp") and xp == nd_array_field.cp:
+            yield pytest.param(xp, id=xp.__name__, marks=pytest.mark.requires_gpu)
+        else:
+            yield pytest.param(xp, id=xp.__name__)
+
+
+@pytest.fixture(params=nd_array_implementation_params())
+def nd_array_implementation(request):
+    yield request.param
