+++ tests/next_tests/unit_tests/embedded_tests/test_nd_array_field.py
@@ -15,7 +15,7 @@ import pytest
 
 from gt4py._core import definitions as core_defs
 from gt4py.next import common
-from gt4py.next.common import Dimension, Domain, UnitRange, NamedRange, NamedIndex
+from gt4py.next.common import Dimension, Domain, NamedIndex, NamedRange, UnitRange
 from gt4py.next.embedded import exceptions as embedded_exceptions, nd_array_field
 from gt4py.next.embedded.nd_array_field import _get_slices_from_domain_slice
 from gt4py.next.ffront import fbuiltins
@@ -26,19 +26,6 @@ from next_tests.integration_tests.feature_tests.math_builtin_test_data import ma
 D0 = Dimension("D0")
 D1 = Dimension("D1")
 D2 = Dimension("D2")
-
-
-def nd_array_implementation_params():
-    for xp in nd_array_field._nd_array_implementations:
-        if hasattr(nd_array_field, "cp") and xp == nd_array_field.cp:
-            yield pytest.param(xp, id=xp.__name__, marks=pytest.mark.requires_gpu)
-        else:
-            yield pytest.param(xp, id=xp.__name__)
-
-
-@pytest.fixture(params=nd_array_implementation_params())
-def nd_array_implementation(request):
-    yield request.param
 
 
 @pytest.fixture(
